+++ types/contracts/cw721-base/approval_response.d.ts
@@ -1,0 +1,5 @@
+import { Approval } from "./shared-types";
+
+export interface ApprovalResponse {
+approval: Approval
+}
