+++ types/contracts/cw721-base/approvals_response.d.ts
@@ -1,0 +1,5 @@
+import { Approval } from "./shared-types";
+
+export interface ApprovalsResponse {
+approvals: Approval[]
+}
