+++ types/contracts/cw721-base/contract_info_response.d.ts
@@ -1,0 +1,4 @@
+export interface ContractInfoResponse {
+name: string
+symbol: string
+}
