+++ types/contracts/cw721-base/minter_response.d.ts
@@ -1,0 +1,6 @@
+/**
+ * Shows who can mint these tokens
+ */
+export interface MinterResponse {
+minter: string
+}
