+++ types/contracts/cw721-base/nft_info_response.d.ts
@@ -1,0 +1,12 @@
+import { Empty } from "./shared-types";
+
+export interface NftInfoResponse {
+/**
+ * You can add any custom metadata here when you extend cw721-base
+ */
+extension?: (Empty | null)
+/**
+ * Universal resource identifier for this NFT Should point to a JSON file that conforms to the ERC721 Metadata JSON Schema
+ */
+token_uri?: (string | null)
+}
