+++ types/contracts/cw721-base/num_tokens_response.d.ts
@@ -1,0 +1,3 @@
+export interface NumTokensResponse {
+count: number
+}
