+++ types/contracts/cw721-base/operators_response.d.ts
@@ -1,0 +1,5 @@
+import { Approval } from "./shared-types";
+
+export interface OperatorsResponse {
+operators: Approval[]
+}
