+++ types/contracts/cw721-base/tokens_response.d.ts
@@ -1,0 +1,6 @@
+export interface TokensResponse {
+/**
+ * Contains all token_ids in lexicographical ordering If there are more than `limit`, use `start_from` in future queries to achieve pagination.
+ */
+tokens: string[]
+}
