+++ utils/package.json
@@ -12,7 +12,8 @@
     "poap": "ts-node poap.ts",
     "cw721-poap": "ts-node cw721-poap.ts",
     "poap-manager": "ts-node poap-manager.ts",
-    "tips": "ts-node tips.ts"
+    "tips": "ts-node tips.ts",
+    "remarkables": "ts-node remarkables.ts"
   },
   "dependencies": {
     "@cosmjs/crypto": "^0.28.11",
